#1 Game Over-popup and High-Scoreboard
Added GameOverPopup Component
In this commit, I've added the GameOverPopup component, which serves as a game over popup. The GameOverPopup component displays the player's name, score, and the cause of death using props. It can be used in various game scenarios.

Added Highscore Component
I've added the Highscore component, designed to display a list of top scores. The Highscore component receives player data and scores through props and dynamically generates list elements using a map function. It's a useful tool for managing and displaying high scores in web applications.